ai_guardian/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
mediapipe==0.10.11
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
ultralytics==8.0.183
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.1
scikit-image==0.21.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.1.1
